substrate_relay_helper

Module cli

Deal with CLI args of substrate-to-substrate relay.

Modules

bridge

Basic traits for exposing bridges in the CLI.

chain_schema

Primitives related to chain CLI options.

detect_equivocations

Primitives for exposing the equivocation detection functionality in the CLI.

init_bridge

Primitives for exposing the bridge initialization functionality in the CLI.

relay_headers

Primitives for exposing the headers relaying functionality in the CLI.

relay_headers_and_messages

Complex 2-ways headers+messages relays support.

relay_messages

Primitives for exposing the messages relaying functionality in the CLI.

relay_parachains

Primitives for exposing the parachains finality relaying functionality in the CLI.

Structs

HexLaneId

Lane id.

PrometheusParams

Prometheus metrics params.

Enums

ExplicitOrMaximal

Either explicit or maximal allowed value.

Constants

LOG_TARGET

The target that will be used when publishing logs related to this pallet.

Type Aliases

DefaultClient

Default Substrate client type that we are using. We’ll use it all over the glue CLI code to avoid multiple level generic arguments and constraints. We still allow usage of other clients in the core logic code.