Module substrate_relay_helper::messages

Tools for supporting message lanes between two Substrate-based chains.

Modules

metrics

Tools for supporting message lanes between two Substrate-based chains.

source

Substrate client as Substrate messages source. The chain we connect to should have runtime that implements <BridgedChainName>HeaderApi to allow bridging with <BridgedName> chain.

target

Substrate client as Substrate messages target. The chain we connect to should have runtime that implements <BridgedChainName>HeaderApi to allow bridging with <BridgedName> chain.

Structs

BatchProofTransaction

Batch transaction that brings headers + and messages delivery/receiving confirmations to the source node.

DirectReceiveMessagesDeliveryProofCallBuilder

Building receive_messages_delivery_proof call when you have direct access to the source chain runtime.

DirectReceiveMessagesProofCallBuilder

Building receive_messages_proof call when you have direct access to the target chain runtime.

MessageLaneAdapter

Adapter that allows all SubstrateMessageLane to act as MessageLane.

MessagesRelayLimits

Delivery transaction limits.

MessagesRelayParams

Substrate <-> Substrate messages relay parameters.

Traits

ReceiveMessagesDeliveryProofCallBuilder

Different ways of building receive_messages_delivery_proof calls.

ReceiveMessagesProofCallBuilder

Different ways of building receive_messages_proof calls.

SubstrateMessageLane

Substrate -> Substrate messages synchronization pipeline.

Functions

relay_messages_delivery_confirmation

Relay messages delivery confirmation of Substrate-to-Substrate messages. No checks are made to ensure that transaction will succeed.

relay_messages_range

Deliver range of Substrate-to-Substrate messages. No checks are made to ensure that transaction will succeed.

run

Run Substrate-to-Substrate messages sync loop.