Crate substrate_relay_helper

The library of substrate relay. contains some public codes to provide to substrate relay.

Modules

cli

Deal with CLI args of substrate-to-substrate relay.

equivocation

Types and functions intended to ease adding of new Substrate -> Substrate equivocation detection pipelines.

error

Relay errors.

finality

Types and functions intended to ease adding of new Substrate -> Substrate finality proofs synchronization pipelines.

finality_base

Types and functions intended to ease adding of new Substrate -> Substrate finality pipelines.

messages

Tools for supporting message lanes between two Substrate-based chains.

on_demand

Types and functions intended to ease adding of new Substrate -> Substrate on-demand pipelines.

parachains

Types and functions intended to ease adding of new Substrate -> Substrate parachain finality proofs synchronization pipelines.

proofs

Module for handling storage proofs compatibility.

Macros

declare_chain_cli_schema

Create chain-specific set of configuration objects: connection parameters, signing parameters and bridge initialization parameters.

declare_chain_connection_params_cli_schema

Create chain-specific set of runtime version parameters.

declare_chain_runtime_version_params_cli_schema

Create chain-specific set of runtime version parameters.

declare_chain_signing_params_cli_schema

Create chain-specific set of signing parameters.

declare_parachain_to_parachain_bridge_schema

Create set of configuration objects specific to parachain-to-parachain relayer.

declare_relay_to_parachain_bridge_schema

Create set of configuration objects specific to relay-to-parachain relayer.

generate_receive_message_delivery_proof_call_builder

Macro that generates ReceiveMessagesDeliveryProofCallBuilder implementation for the case when you only have an access to the mocked version of source chain runtime. In this case you should provide “name” of the call variant for the bridge messages calls and the “name” of the variant for the receive_messages_delivery_proof call within that first option.

generate_receive_message_proof_call_builder

Macro that generates ReceiveMessagesProofCallBuilder implementation for the case when you only have an access to the mocked version of target chain runtime. In this case you should provide “name” of the call variant for the bridge messages calls and the “name” of the variant for the receive_messages_proof call within that first option.

generate_report_equivocation_call_builder

Macro that generates ReportEquivocationCallBuilder implementation for the case where we only have access to the mocked version of the source chain runtime.

generate_submit_finality_proof_call_builder

Macro that generates SubmitFinalityProofCallBuilder implementation for the case when you only have an access to the mocked version of target chain runtime. In this case you should provide “name” of the call variant for the bridge GRANDPA calls and the “name” of the variant for the submit_finality_proof call within that first option.

generate_submit_finality_proof_ex_call_builder

Macro that generates SubmitFinalityProofCallBuilder implementation for the case when you only have an access to the mocked version of target chain runtime. In this case you should provide “name” of the call variant for the bridge GRANDPA calls and the “name” of the variant for the submit_finality_proof_ex call within that first option.

Structs

TransactionParams

Transaction creation parameters.

UtilityPalletBatchCallBuilder

Batch call builder based on pallet-utility.

Enums

HeadersToRelay

Type of headers that we relay.

TaggedAccount

Tagged relay account, which balance may be exposed as metrics by the relay.

Traits

BatchCallBuilder

Batch call builder.

BatchCallBuilderConstructor

Batch call builder constructor.